# crypto: Performance Improvement for `Keccak256` and `Keccak256Hash`