This PR improves the performance of `Keccak256` and `Keccak256Hash` by adding a `sync.Pool` to reuse `KeccakState`'s. 

```
name    old time/op  new time/op  delta
Sha3-8  1.29µs ± 5%  0.73µs ± 5%  -42.95%  (p=0.000 n=20+20)
```
As `crypto.Keccak256` is widely used there are likely performance improvements across the codebase. The EVM execution of a simple WETH9 transfer is ~3% faster.