@lmittmann 2022-06-11
A `sync.Pool` for `KeccakState`'s is already used in https://github.com/ethereum/go-ethereum/blob/master/core/types/hashing.go#L30. It should probably be removed in favor of the global pool.