@holiman 2022-06-11
Please see discussion at https://github.com/ethereum/go-ethereum/pull/24780