@lmittmann 2022-06-11
@holiman did not see that. In that case reuse of `KeccakState` should be done in the EVM tough, because this change improves EVM execution speed.