@holiman 2022-06-13
>  reuse of KeccakState should be done in the EVM though, because this change improves EVM execution speed.

If you think that would be a meaningful change, please file a PR for that. I personally am not yet convinced that it will be -- but that depends also on how often keccak256 is invoked. If the cache is per-tx and we only invoke it a couple of times per block, then it's probably not worth it. 

Closing this PR, for the same reasons as given in #24780